Added all necessary methods except getEmployees
After two days i've been trying to solve this Case Study, finally i have figured out how to finish all the methods as of what the requirement's needed.

BUT, i still cannot figured out how to create a method when it's invoked, it'll display a list of employee in that Department. Or i can say getEmployees method.

Now i'm working on it for how to fix it.
More to come.