Temporary solutins for getEmployees
Finally i can add a functional getEmployees() method.
The Problem is when i invoke the method it successfully show a list of employees of that Department.

However, it always printed null at the end of the list.
For example =
HRApp Starts
Department name = IT Support Department
Total employee in this department = 10
Employee list =
123, Putra Ganda, 13.5
124, Josh, 12.5
125, Steve, 14.0
126, Arash, 14.0
127, Ali, 14.0
128, Yousef, 14.0
129, Jack, 14.0
130, Casey, 14.0
131, Chloe, 14.0
132, Cassandra, 14.0
null

Total Salary = 138.0
Average Salary = 13.8

This is just a temporary fix until i found a better approach/solution to make getEmployees method more perfect. More to come.